+++ sentry-android-core/src/main/java/io/sentry/android/core/ManifestMetadataReader.java
@@ -55,6 +55,8 @@ final class ManifestMetadataReader {
   static final String TRACES_ACTIVITY_AUTO_FINISH_ENABLE =
       "io.sentry.traces.activity.auto-finish.enable";
 
+  static final String TRACES_PROFILING_ENABLE = "io.sentry.traces.profiling.enable";
+
   @ApiStatus.Experimental static final String TRACE_SAMPLING = "io.sentry.traces.trace-sampling";
 
   static final String TRACING_ORIGINS = "io.sentry.traces.tracing-origins";
@@ -222,6 +224,9 @@ final class ManifestMetadataReader {
                 logger,
                 TRACES_ACTIVITY_AUTO_FINISH_ENABLE,
                 options.isEnableActivityLifecycleTracingAutoFinish()));
+
+        options.setProfilingEnabled(
+            readBool(metadata, logger, TRACES_PROFILING_ENABLE, options.isProfilingEnabled()));
 
         final List<String> tracingOrigins = readList(metadata, logger, TRACING_ORIGINS);
         if (tracingOrigins != null) {