+++ sentry-android-core/src/main/java/io/sentry/android/core/SentryAndroidOptions.java
@@ -7,6 +7,7 @@ import io.sentry.SentryOptions;
 import io.sentry.SpanStatus;
 import io.sentry.protocol.SdkVersion;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 /** Sentry SDK options for Android */
 public final class SentryAndroidOptions extends SentryOptions {
@@ -83,6 +84,12 @@ public final class SentryAndroidOptions extends SentryOptions {
    * Spans.
    */
   private boolean enableActivityLifecycleTracingAutoFinish = true;
+
+  /** The cache dir. path for caching profiling traces */
+  private @Nullable String profilingTracesDirPath;
+
+  /** Interval for profiling traces in milliseconds. Defaults to 300 times per second */
+  private int profilingTracesIntervalMillis = 1_000 / 300;
 
   /** Interface that loads the debug images list */
   private @NotNull IDebugImagesLoader debugImagesLoader = NoOpDebugImagesLoader.getInstance();
@@ -217,6 +224,42 @@ public final class SentryAndroidOptions extends SentryOptions {
     enableSystemEventBreadcrumbs = enable;
     enableAppLifecycleBreadcrumbs = enable;
     enableUserInteractionBreadcrumbs = enable;
+  }
+
+  /**
+   * Returns the profiling traces dir. path if set
+   *
+   * @return the profiling traces dir. path or null if not set
+   */
+  public @Nullable String getProfilingTracesDirPath() {
+    return profilingTracesDirPath;
+  }
+
+  /**
+   * Sets the profiling traces dir. path
+   *
+   * @param profilingTracesDirPath the profiling traces dir. path
+   */
+  public void setProfilingTracesDirPath(@Nullable String profilingTracesDirPath) {
+    this.profilingTracesDirPath = profilingTracesDirPath;
+  }
+
+  /**
+   * Returns the interval for profiling traces in milliseconds.
+   *
+   * @return the interval for profiling traces in milliseconds.
+   */
+  public int getProfilingTracesIntervalMillis() {
+    return profilingTracesIntervalMillis;
+  }
+
+  /**
+   * Sets the interval for profiling traces in milliseconds.
+   *
+   * @param profilingTracesIntervalMillis - the interval for profiling traces in milliseconds.
+   */
+  public void setProfilingTracesIntervalMillis(final int profilingTracesIntervalMillis) {
+    this.profilingTracesIntervalMillis = profilingTracesIntervalMillis;
   }
 
   /**