+++ sentry-android-core/src/test/java/io/sentry/android/core/internal/gestures/SentryGestureListenerScrollTest.kt
@@ -22,8 +22,8 @@ import io.sentry.Breadcrumb
 import io.sentry.IHub
 import io.sentry.SentryLevel.INFO
 import io.sentry.android.core.SentryAndroidOptions
-import org.junit.Test
 import java.lang.ref.WeakReference
+import kotlin.test.Test
 import kotlin.test.assertEquals
 
 class SentryGestureListenerScrollTest {