+++ sentry-apollo/src/test/java/io/sentry/apollo/SentryApolloInterceptorTest.kt
@@ -85,7 +85,9 @@ class SentryApolloInterceptorTest {
                 assertTransactionDetails(it)
                 assertEquals(SpanStatus.OK, it.spans.first().status)
             },
-            anyOrNull<TraceState>()
+            anyOrNull<TraceState>(),
+            anyOrNull(),
+            anyOrNull()
         )
     }
 
@@ -98,7 +100,9 @@ class SentryApolloInterceptorTest {
                 assertTransactionDetails(it)
                 assertEquals(SpanStatus.PERMISSION_DENIED, it.spans.first().status)
             },
-            anyOrNull<TraceState>()
+            anyOrNull<TraceState>(),
+            anyOrNull(),
+            anyOrNull()
         )
     }
 
@@ -111,7 +115,9 @@ class SentryApolloInterceptorTest {
                 assertTransactionDetails(it)
                 assertEquals(SpanStatus.INTERNAL_ERROR, it.spans.first().status)
             },
-            anyOrNull<TraceState>()
+            anyOrNull<TraceState>(),
+            anyOrNull(),
+            anyOrNull()
         )
     }
 
@@ -145,7 +151,9 @@ class SentryApolloInterceptorTest {
                 val httpClientSpan = it.spans.first()
                 assertEquals("overwritten description", httpClientSpan.description)
             },
-            anyOrNull<TraceState>()
+            anyOrNull<TraceState>(),
+            anyOrNull(),
+            anyOrNull()
         )
     }
 
@@ -159,7 +167,9 @@ class SentryApolloInterceptorTest {
             check {
                 assertEquals(1, it.spans.size)
             },
-            anyOrNull<TraceState>()
+            anyOrNull<TraceState>(),
+            anyOrNull(),
+            anyOrNull()
         )
     }