+++ sentry/src/main/java/io/sentry/NoOpSentryClient.java
@@ -50,7 +50,8 @@ final class NoOpSentryClient implements ISentryClient {
       @NotNull SentryTransaction transaction,
       @Nullable TraceState traceState,
       @Nullable Scope scope,
-      @Nullable Map<String, Object> hint) {
+      @Nullable Map<String, Object> hint,
+      @Nullable ProfilingTraceData profilingTraceData) {
     return SentryId.EMPTY_ID;
   }
 }