+++ sentry/src/main/java/io/sentry/SentryClient.java
@@ -2,6 +2,7 @@ package io.sentry;
 
 import static io.sentry.TypeCheckHint.SENTRY_SCREENSHOT;
 
+import io.sentry.exception.SentryEnvelopeException;
 import io.sentry.hints.DiskFlushNotification;
 import io.sentry.protocol.Contexts;
 import io.sentry.protocol.SentryId;
@@ -139,12 +140,12 @@ public final class SentryClient implements ISentryClient {
               ? scope.getTransaction().traceState()
               : null;
       final SentryEnvelope envelope =
-          buildEnvelope(event, getAttachments(scope, hint), session, traceState);
+          buildEnvelope(event, getAttachments(scope, hint), session, traceState, null);
 
       if (envelope != null) {
         transport.send(envelope, hint);
       }
-    } catch (IOException e) {
+    } catch (IOException | SentryEnvelopeException e) {
       options.getLogger().log(SentryLevel.WARNING, e, "Capturing event %s failed.", sentryId);
 
       // if there was an error capturing the event, we return an emptyId
@@ -178,8 +179,9 @@ public final class SentryClient implements ISentryClient {
       final @Nullable SentryBaseEvent event,
       final @Nullable List<Attachment> attachments,
       final @Nullable Session session,
-      final @Nullable TraceState traceState)
-      throws IOException {
+      final @Nullable TraceState traceState,
+      final @Nullable ProfilingTraceData profilingTraceData)
+      throws IOException, SentryEnvelopeException {
     SentryId sentryId = null;
 
     final List<SentryEnvelopeItem> envelopeItems = new ArrayList<>();
@@ -195,6 +197,13 @@ public final class SentryClient implements ISentryClient {
       final SentryEnvelopeItem sessionItem =
           SentryEnvelopeItem.fromSession(options.getSerializer(), session);
       envelopeItems.add(sessionItem);
+    }
+
+    if (profilingTraceData != null) {
+      final SentryEnvelopeItem profilingTraceItem =
+          SentryEnvelopeItem.fromProfilingTrace(
+              profilingTraceData, options.getMaxTraceFileSize(), options.getSerializer());
+      envelopeItems.add(profilingTraceItem);
     }
 
     if (attachments != null) {
@@ -423,7 +432,8 @@ public final class SentryClient implements ISentryClient {
       @NotNull SentryTransaction transaction,
       @Nullable TraceState traceState,
       final @Nullable Scope scope,
-      @Nullable Map<String, Object> hint) {
+      @Nullable Map<String, Object> hint,
+      final @Nullable ProfilingTraceData profilingTraceData) {
     Objects.requireNonNull(transaction, "Transaction is required.");
 
     if (hint == null) {
@@ -463,13 +473,18 @@ public final class SentryClient implements ISentryClient {
     try {
       final SentryEnvelope envelope =
           buildEnvelope(
-              transaction, filterForTransaction(getAttachments(scope, hint)), null, traceState);
+              transaction,
+              filterForTransaction(getAttachments(scope, hint)),
+              null,
+              traceState,
+              profilingTraceData);
+
       if (envelope != null) {
         transport.send(envelope, hint);
       } else {
         sentryId = SentryId.EMPTY_ID;
       }
-    } catch (IOException e) {
+    } catch (IOException | SentryEnvelopeException e) {
       options.getLogger().log(SentryLevel.WARNING, e, "Capturing transaction %s failed.", sentryId);
       // if there was an error capturing the event, we return an emptyId
       sentryId = SentryId.EMPTY_ID;