+++ sentry/src/main/java/io/sentry/SentryItemType.java
@@ -13,6 +13,7 @@ public enum SentryItemType implements JsonSerializable {
   UserFeedback("user_report"), // Sentry backend still uses user_report
   Attachment("attachment"),
   Transaction("transaction"),
+  Profile("profile"),
   Unknown("__unknown__"); // DataCategory.Unknown
 
   private final String itemType;