+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 * Feat: Screenshot is taken when there is an error (#1967)
+* Feat: Add Android profiling traces #1897 and its tests #1949
+- All operations involving file reads for profiling were moved to the background #1959
 
 ## 6.0.0-alpha.4